CLN, ENH: update logging system. Reorganize callbacks
