+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [4.2.1] - 2022-08-14
+
+## Changed
+- Logging now is handled by `accelerate` library.
+- Callbacks have been reorganized in folders.
+- Update requirements.
+- Update `gh-pages.yml`.
+- Update `README.md`
+- Update `mixed_precision` parameter and usage.
+
+
 ## [4.2.0] - 2022-08-08
 
 ## Added
