+++ requirements-dev.txt
@@ -1,9 +1,9 @@
 -r requirements.txt
 -r requirements-doc.txt
 
-pytest==5.4.3
-black==20.8b1
-mypy==0.942
-mypy-extensions==0.4.3
-flake8==4.0.1
-isort==5.10.1
+pytest
+black
+mypy
+mypy-extensions
+flake8
+isort
