+++ requirements-doc.txt
@@ -1,9 +1,10 @@
-Sphinx==4.1.2
-sphinx-autobuild==2021.3.14
-sphinxcontrib-applehelp==1.0.2
-sphinxcontrib-devhelp==1.0.2
-sphinxcontrib-htmlhelp==2.0.0
+sphinx-rtd-theme==0.5.1
+sphinxcontrib-applehelp==1.0.1
+sphinxcontrib-devhelp==1.0.1
+sphinxcontrib-htmlhelp==1.0.2
 sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==1.0.3
-sphinxcontrib-serializinghtml==1.1.5
-furo==2021.7.31b41
+sphinxcontrib-qthelp==1.0.2
+sphinxcontrib-serializinghtml==1.1.3
+sphinxcontrib-websupport==1.1.2
+furo==2020.12.30b24
+Jinja2==3.0.3  # see https://github.com/sphinx-doc/sphinx/issues/10291
