+++ requirements.txt
@@ -1,4 +1,4 @@
-accelerate==0.11.0
+accelerate==0.12.0
 certifi==2020.12.5
 joblib==1.0.1
 numpy==1.20.3
@@ -7,6 +7,6 @@ scipy==1.5.1
 threadpoolctl==2.1.0
 torch==1.8.1
 torchmetrics==0.5.0
-typing-extensions==3.7.4.3
+typing-extensions==3.10.0.0
 wincertstore==0.2
 rich==10.14.0
