+++ src/torchfitter/callbacks/__init__.py
@@ -1,16 +1,18 @@
 """ Abstractions to interact with models. """
 
+from torchfitter.callbacks.progress import LoggerCallback, RichProgressBar
+from torchfitter.callbacks.regularization import (
+    ElasticNetRegularization,
+    L1Regularization,
+    L2Regularization,
+)
+
 from . import base
 from ._callbacks import (
     EarlyStopping,
     GPUStats,
     LearningRateScheduler,
-    LoggerCallback,
-    RichProgressBar,
     StochasticWeightAveraging,
-    L1Regularization,
-    L2Regularization,
-    ElasticNetRegularization,
 )
 
 __all__ = [
