+++ src/torchfitter/callbacks/_callbacks.py
@@ -3,11 +3,10 @@ import subprocess
 from typing import List
 
 import torch
-from rich.progress import BarColumn, Progress, TimeRemainingColumn
 from torch.optim.swa_utils import AveragedModel
 
-from torchfitter.conventions import ParamsDict
 from torchfitter.callbacks.base import Callback
+from torchfitter.conventions import ParamsDict
 
 
 class EarlyStopping(Callback):
@@ -33,7 +32,8 @@ class EarlyStopping(Callback):
         self.patience = patience
         self.load_best = load_best
 
-        self.log_name = "EarlyStopping"
+        self.set_log_name("EarlyStopping")
+        self.set_log_level(20)  # info logging
 
     def __repr__(self) -> str:
         return (
@@ -75,68 +75,6 @@ class EarlyStopping(Callback):
             self.logger.info(
                 f"Early stopping applied at epoch: {self.stopped_epoch}"
             )
-
-
-class LoggerCallback(Callback):
-    """
-    `LoggerCallback` is used to log some of the parameters dict in order to
-    monitor the fitting process. The logged parameters are:
-        - Current epoch.
-        - Number of epochs.
-        - Train loss.
-        - Validation loss.
-        - Time / epoch.
-
-    It also outputs the total training time once the fitting process ends.
-
-    Parameters
-    ----------
-    update_step : int, optional, default: 50
-        Logs will be performed every 'update_step'.
-    precision : int, optional, default: 2
-        Number of decimals in the numbers.
-    """
-
-    def __init__(self, update_step: int, precision: int = 2):
-        super(LoggerCallback, self).__init__()
-        self.update_step = update_step
-        self.prec = precision
-
-        self.log_name = "LoggerCallback"
-
-    def __repr__(self) -> str:
-        return (
-            f"LoggerCallback(update_step={self.update_step}, "
-            f"precision={self.prec})"
-        )
-
-    def on_fit_start(self, params_dict: dict) -> None:
-        dev = params_dict[ParamsDict.DEVICE]
-        self.logger.info(f"Starting training process on {dev}")
-
-    def on_epoch_end(self, params_dict: dict) -> None:
-        epoch_number = params_dict[ParamsDict.EPOCH_NUMBER]
-        total_epochs = params_dict[ParamsDict.TOTAL_EPOCHS]
-        val_loss = params_dict[ParamsDict.VAL_LOSS]
-        train_loss = params_dict[ParamsDict.TRAIN_LOSS]
-        epoch_time = params_dict[ParamsDict.EPOCH_TIME]
-
-        prec = self.prec
-        msg = (
-            f"Epoch {epoch_number}/{total_epochs} | Train loss: "
-            f"{train_loss:.{prec}e} | Validation loss {val_loss:.{prec}e} | "
-            f"Time/epoch: {epoch_time:.{prec}e} s"
-        )
-
-        if epoch_number % self.update_step == 0 or epoch_number == 1:
-            self.logger.info(msg)
-
-    def on_fit_end(self, params_dict: dict) -> None:
-        total_time = params_dict[ParamsDict.TOTAL_TIME]
-        # final message
-        self.logger.info(
-            f"""End of training. Total time: {total_time:0.5f} seconds"""
-        )
 
 
 class LearningRateScheduler(Callback):
@@ -264,7 +202,8 @@ class GPUStats(Callback):
         self.format = format
         self.update_step = update_step
 
-        self.log_name = "GPU Stats"
+        self.set_log_name("GPU Stats")
+        self.set_log_level(20)  # info logging
 
     def __repr__(self) -> str:
         return (
@@ -299,134 +238,6 @@ class GPUStats(Callback):
             stdout.append(out)
 
         return stdout
-
-
-class RichProgressBar(Callback):
-    """
-    This callback displays a progress bar to report the state of the training
-    process: on each epoch, a new bar will be created and stacked below the
-    previous bars.
-
-    Metrics are logged using the library logger.
-
-    Parameters
-    ----------
-    display_step : int
-        Number of epochs to wait to display the progress bar.
-    precision : int, optional, default: 2
-        Number of decimals to use in the log.
-    log_lr : bool, optional, default: False
-        Whether to log the learning rate (True) or not (False).
-    """
-
-    def __init__(
-        self, display_step: int = 1, log_lr: bool = False, precision: int = 2
-    ):
-        super(RichProgressBar, self).__init__()
-
-        self.display_step = display_step
-        self.prec = precision
-        self.log_lr = log_lr
-
-        self.log_name = "Rich Bar"
-
-    def __repr__(self) -> str:
-        return (
-            f"RichProgressBar(display_step={self.display_step}, "
-            f"log_lr={self.log_lr}, precision={self.precision})"
-        )
-
-    def on_fit_start(self, params_dict: dict) -> None:
-        dev = params_dict[ParamsDict.DEVICE]
-        self.logger.info(f"Starting training process on {dev}\n")
-
-    def on_train_batch_end(self, params_dict: dict) -> None:
-        epoch = params_dict[ParamsDict.EPOCH_NUMBER]
-        accelerator = params_dict[ParamsDict.ACCELERATOR]
-
-        if epoch % self.display_step == 0 or epoch == 1:
-            accelerator.wait_for_everyone()
-            self.progress_bar.advance(self.epoch_task, 1)
-
-    def on_validation_batch_end(self, params_dict: dict) -> None:
-        epoch = params_dict[ParamsDict.EPOCH_NUMBER]
-        accelerator = params_dict[ParamsDict.ACCELERATOR]
-
-        if epoch % self.display_step == 0 or epoch == 1:
-            accelerator.wait_for_everyone()
-            self.progress_bar.advance(self.epoch_task, 1)
-
-    def on_epoch_start(self, params_dict: dict) -> None:
-        # gather necessary objects
-        train_loader = params_dict[ParamsDict.TRAIN_LOADER]
-        val_loader = params_dict[ParamsDict.VAL_LOADER]
-        epoch = params_dict[ParamsDict.EPOCH_NUMBER]
-        total_epochs = params_dict[ParamsDict.TOTAL_EPOCHS]
-
-        # compute number of batches
-        n_elements = len(train_loader) + len(val_loader)
-
-        if epoch % self.display_step == 0 or epoch == 1:
-            self.progress_bar = Progress(
-                "[progress.description]{task.description}",
-                "•",
-                BarColumn(),
-                "•",
-                "[progress.percentage]{task.percentage:>3.0f}%",
-                "•",
-                TimeRemainingColumn(),
-            )
-
-            self.epoch_task = self.progress_bar.add_task(
-                description=f"Epoch {epoch}/{total_epochs}",
-                total=n_elements,
-            )
-            self.progress_bar.start()
-
-    def on_epoch_end(self, params_dict: dict) -> None:
-        epoch = params_dict[ParamsDict.EPOCH_NUMBER]
-
-        if epoch % self.display_step == 0 or epoch == 1:
-            # update metrics
-            text = self.render_text(params_dict[ParamsDict.EPOCH_HISTORY])
-            self.logger.info(text)  # DISC: use included Rich logger?
-            self.progress_bar.stop()
-
-    def on_fit_end(self, params_dict):
-        total_time = params_dict[ParamsDict.TOTAL_TIME]
-        # final message
-        self.logger.info(
-            f"""\nEnd of training. Total time: {total_time:0.5f} seconds"""
-        )
-
-    def render_text(self, update_dict):
-        text_format = ""
-
-        for metric in update_dict:
-            if metric != ParamsDict.HISTORY_LR:
-                train_metric = update_dict[metric]["train"][-1]
-                val_metric = update_dict[metric]["validation"][-1]
-
-                if text_format:  # not empty
-                    text_format = (
-                        f"{text_format} • {metric} -> Train: "
-                        f"{train_metric:.{self.prec}e} | "
-                        f"Validation: {val_metric:.{self.prec}e}"
-                    )
-                else:
-                    text_format = (
-                        f"{metric} -> Train: "
-                        f"{train_metric:.{self.prec}e} | Validation: "
-                        f"{val_metric:.{self.prec}e}"
-                    )
-            else:
-                if self.log_lr:
-                    text_format = (
-                        f"{text_format} • Learning Rate: "
-                        f"{update_dict[metric][-1]}"
-                    )
-
-        return text_format
 
 
 class StochasticWeightAveraging(Callback):
@@ -549,164 +360,3 @@ class StochasticWeightAveraging(Callback):
             SWA model.
         """
         return self.__swa_model
-
-
-class L1Regularization(Callback):
-    """Applies L1 regularization over the model parameters.
-
-    L1 is usually called 'Lasso Regression' (Least Absolute Shrinkage and
-    Selection Operator). This callbacks is only applied to the train loss.
-
-    Parameters
-    ----------
-    regularization_rate : float
-        Regularization rate. Also called `lambda`.
-    biases : bool, optional, default: False
-        Whether to apply regularization over bias terms (True) or not (False).
-
-    Note
-    ----
-    The penalty term already handles the product by the lambda regularization
-    rate.
-
-    """
-
-    def __init__(self, regularization_rate: float, biases: bool = False):
-        super().__init__()
-
-        self.rate = regularization_rate
-        self.biases = biases
-
-    def on_loss_step_end(self, params_dict: dict) -> None:
-        batch_tr_loss = params_dict[ParamsDict.BATCH_TRAIN_LOSS]
-        device = params_dict[ParamsDict.DEVICE]
-        model = params_dict[ParamsDict.MODEL]
-
-        # Initialize with tensor, cannot be scalar
-        penalty_term = torch.zeros(1, 1, requires_grad=True).to(device)
-
-        for name, param in model.named_parameters():
-            if not self.biases and name.endswith("bias"):
-                continue
-
-            penalty_term = penalty_term + param.norm(p=1)
-
-        total_penalty = self.rate * penalty_term
-        loss = total_penalty + batch_tr_loss
-
-        # set loss
-        params_dict[ParamsDict.BATCH_TRAIN_LOSS] = loss
-
-
-class L2Regularization(Callback):
-    """Applies L2 regularization over the model parameters.
-
-    L2 is usually called 'Ridge Regression'. This callbacks is only applied to
-    the train loss.
-
-    Parameters
-    ----------
-    regularization_rate : float
-        Regularization rate. Also called `lambda`.
-    biases : bool, optional, default: False
-        Whether to apply regularization over bias terms (True) or not (False).
-
-    Note
-    ----
-    The penalty term already handles the product by the lambda regularization
-    rate.
-
-    """
-
-    def __init__(self, regularization_rate: float, biases: bool = False):
-        super().__init__()
-
-        self.rate = regularization_rate
-        self.biases = biases
-
-    def on_loss_step_end(self, params_dict: dict) -> None:
-        batch_tr_loss = params_dict[ParamsDict.BATCH_TRAIN_LOSS]
-        device = params_dict[ParamsDict.DEVICE]
-        model = params_dict[ParamsDict.MODEL]
-
-        # Initialize with tensor, cannot be scalar
-        penalty_term = torch.zeros(1, 1, requires_grad=True).to(device)
-
-        for name, param in model.named_parameters():
-            if not self.biases and name.endswith("bias"):
-                continue
-
-            penalty_term = penalty_term + param.norm(p=2)
-
-        total_penalty = self.rate * penalty_term
-        loss = total_penalty + batch_tr_loss
-
-        # set loss
-        params_dict[ParamsDict.BATCH_TRAIN_LOSS] = loss
-
-
-class ElasticNetRegularization(Callback):
-    r"""Linear combination of L1 and L2.
-
-    According to [1], the lasso penalty is somewhat indifferent to the choice
-    among a set of strong but correlated variables. The ridge penalty, on the
-    other hand, tends to shrink the coefficients of correlated variables toward
-    each other. Elastic net combines both using a weighting factor:
-
-    .. math::
-
-        \sum_{j=1}^{p} ( \alpha |\beta_{j}| + (1 + \alpha) \beta_{j}^{2} )
-
-    Parameters
-    ----------
-    regularization_rate : float
-        Regularization rate. Also called `lambda`.
-    alpha : float
-        Parameter to determine the mix of the penalties.
-    biases : bool, optional, default: False
-        Whether to apply regularization over bias terms (True) or not (False).
-
-    Note
-    ----
-    The penalty term already handles the product by the lambda regularization
-    rate.
-
-    References
-    ----------
-    .. [1] Trevor Hastie, Robert Tibshirani, Jerome Friedman - The Elements of
-       Statistical Learning.
-
-    """
-
-    def __init__(
-        self, regularization_rate: float, alpha: float, biases: bool = False
-    ):
-        super().__init__()
-
-        self.rate = regularization_rate
-        self.biases = biases
-        self.alpha = alpha
-
-    def on_loss_step_end(self, params_dict: dict) -> None:
-        batch_tr_loss = params_dict[ParamsDict.BATCH_TRAIN_LOSS]
-        device = params_dict[ParamsDict.DEVICE]
-        model = params_dict[ParamsDict.MODEL]
-
-        # Initialize with tensor, cannot be scalar
-        penalty_term = torch.zeros(1, 1, requires_grad=True).to(device)
-
-        for name, param in model.named_parameters():
-            if not self.biases and name.endswith("bias"):
-                continue
-
-            l1 = param.norm(p=1)
-            l2 = param.norm(p=2)
-            penalty_term = penalty_term + (
-                self.alpha * l1 + (1 - self.alpha) * l2
-            )
-
-        total_penalty = self.rate * penalty_term
-        loss = total_penalty + batch_tr_loss
-
-        # set loss
-        params_dict[ParamsDict.BATCH_TRAIN_LOSS] = loss
