+++ src/torchfitter/trainer/_trainer.py
@@ -1,4 +1,3 @@
-import logging
 import statistics
 import time
 from typing import List, Tuple, Union
@@ -26,6 +25,10 @@ class Trainer:
     at different points depending on the methods that were filled. The metrics
     will be runned in the train and validation steps.
 
+    The callbacks will run in the passed order. Meaning, if a callback modifies
+    the loss value after the logging has been performed, the new loss value
+    won't be showed in the logging process.
+
     Parameters
     ----------
     model : torch.Module
@@ -34,11 +37,13 @@ class Trainer:
         Loss function criterion used to optimize the model.
     optimizer : torch.optim
         Optimizer to perform the parameters update.
-    mixed_precision : bool, optional, default: False
-        Whether to use mixed precision training or not. If True, the forward
-        pass will be computed under the context of `torch.cuda.amp.autocast`
-        and the backpropagation and gradient descent steps will be computed
-        using `torch.cuda.amp.GradScaler`.
+    mixed_precision : str, optional, {'no', 'fp16', 'bf16'}, default: None
+        Whether or not to use mixed precision training (fp16 or bfloat16). From
+        Accelerate DOC: choose from 'no', 'fp16', 'bf16'. Will default to the
+        value in the environment variable MIXED_PRECISION, which will use the
+        default value in the accelerate config of the current system or the
+        flag passed with the accelerate.launch command. 'fp16' requires pytorch
+        1.6 or higher. 'bf16' requires pytorch 1.10 or higher.
     callbacks : list of torchfitter.callback.Callback
         Callbacks to use during the training process.
     metrics : list of torchmetrics.Metric, optional, default: None
@@ -61,8 +66,6 @@ class Trainer:
         algorithm. Example: {max_norm=1, norm_type=2}. See
         https://huggingface.co/docs/accelerate/accelerator.html for more
         information.
-    log_level : int, optional, default: logging.INFO
-        Logging level.
 
     Attributes
     ----------
@@ -75,6 +78,9 @@ class Trainer:
     gradient_clipping_algo_ : callable
         Gradient clipping algorithm or None.
 
+    TODO
+    ----
+    - Make callbacks with priority and sort them at the beginning.
     """
 
     def __init__(
@@ -82,14 +88,13 @@ class Trainer:
         model: torch.nn.Module,
         criterion: torch.nn.Module,
         optimizer: torch.optim.Optimizer,
-        mixed_precision: bool = False,
+        mixed_precision: str = None,
         callbacks: List[Callback] = None,
         metrics: List[torchmetrics.Metric] = None,
         accelerator: Accelerator = None,
         accumulate_iter: int = 1,
         gradient_clipping: str = None,
         gradient_clipping_kwargs: dict = None,
-        log_level: int = logging.INFO,
     ):
         self.criterion = criterion
         self.callbacks_list = callbacks
@@ -97,12 +102,12 @@ class Trainer:
         self.accumulate_iter = accumulate_iter
         self.gradient_clipping = gradient_clipping
         self.gradient_clipping_kwargs = gradient_clipping_kwargs
-        self.log_level = log_level
 
         if accelerator is None:
             self.accelerator = Accelerator(
                 fp16=mixed_precision,
                 gradient_accumulation_steps=accumulate_iter,
+                step_scheduler_with_optimizer=True,
             )
 
         # wrap withing accelerate environment
@@ -116,7 +121,6 @@ class Trainer:
         self.callback_handler = CallbackHandler(
             callbacks_list=self.callbacks_list
         )
-        self.callback_handler.set_log_level(self.log_level)
 
         self.metrics_handler = MetricsHandler(
             metrics_list=self.metrics_list,
@@ -334,23 +338,6 @@ class Trainer:
                 f"gradient clipping algorithm: '{self.gradient_clipping}'"
             )
         return algo
-
-    def set_scaler(
-        self, scaler: torch.cuda.amp.grad_scaler.GradScaler
-    ) -> None:
-        """Set the gradient scaler used in mixed precision.
-
-        Since the trainer relies on accelerate.Accelator class for the fitting
-        process the scaler is created inside the aforementioned. With this
-        function you can set the scaler to be an instance with the desired
-        argument values.
-
-        Parameters
-        ----------
-        scaler : torch.cuda.amp.grad_scaler.GradScaler
-            The gradient scaler you want to use.
-        """
-        self.accelerator.scaler = scaler
 
     def reset_parameters(self, reset_model: bool = False) -> None:
         """
